refactor(data): remove api.ts — move interfaces into adapters/index.ts
