+++ packages/scope42-data/src/io/adapters/node.ts
@@ -3,7 +3,7 @@
  * such as the linter CLI, tests running outside a browser, and the example
  * in `examples/data-processing`.
  */
-import { DirectoryHandle, FileHandle } from './api'
+import { DirectoryHandle, FileHandle } from '.'
 import { promisify } from 'util'
 import fs from 'fs'
 import { basename } from 'path'
